pyproject.toml
[project]
name = "deeplearning-mcp-tutorial"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "arxiv==2.2.0",
    "anthropic==0.51.0",
    "dotenv==0.9.9",
    "fastmcp",
    "nest-asyncio"
]
